feat(global): console 输出emoji && 输入内容两种交互

+++ command/publish.js
@@ -1,33 +1,41 @@
 'use strict'
 const co = require('co')
 const prompt = require('co-prompt')
-// const config = require('../templates')
-const chalk = require('chalk')
-const fs = require('fs')
-const { exec } = require('child_process');
+const logger = require('../src/logger')
 const path = require('path');
+const fs = require('fs')
+const exists = require('fs').existsSync
+const { exec } = require('child_process');
+
 const getGitUser = require('../src/git-user')
 const operFile = require('../src/oper-file')
 const execPromise = require('../src/exec-promise')
 
 const author = getGitUser()
-console.log('author', author.name)
 
 module.exports = () => {
   co(function* () {
-    // 分步接收用户输入的参数
-    let packageName = yield prompt('Package name: ')
+    let [a, b, target, ...params] = process.argv
+    let packageName = params[0]
 
-    console.log('process.cwd()', process.cwd())
+    if (!packageName) {
+      packageName = yield prompt('Package name: ')
+    }
+
     let filename = process
       .cwd()
       .replace(/[\\]/g, '/') + '/' + packageName
 
     let templateSrc = path.join(__dirname, '../template')
 
-    // todo 判断文件夹是否存在
-    const cmd = `mkdir ${packageName} && cp ${templateSrc}/* ${filename}`
-    console.log('cmd', cmd)
+    let cmd
+    if (exists(packageName)) {
+      cmd = `cp ${templateSrc}/* ${filename}`
+      logger.info('已存在该目录， 请更换目录或包名之后重试')
+      process.exit()
+    } else {
+      cmd = `mkdir ${packageName} && cp ${templateSrc}/* ${filename}`
+    }
 
     execPromise(cmd).then(resp => {
 
@@ -41,41 +49,36 @@ module.exports = () => {
           email: author.email,
           name: packageName
         }
-        // console.log('newData', newData)
+
         operFile.writeFilePromise(packageJsonSrc, newData).then(data => {
-          console.log('写入成功')
-          console.log('退出进程。。。')
+          logger.success('写入成功')
+          logger.info('退出进程。。。')
         }, err => {
-          console.log('写入失败')
-          console.log('退出进程。。。')
+          logger.fail('写入失败')
+          logger.info('退出进程。。。')
           process.exit()
         })
 
       }, err => {
-        console.log('err', err)
-        console.log('退出进程。。。')
+        logger.fail('err' + err)
+        logger.info('退出进程。。。')
         process.exit()
       })
 
 
       // publish 发布npm包
       // 如果没有登录npm 需要手动先登录npm。
-      // 通过代码来自动获取 username
-      // todo console 输出 emoji
-      // && npm publish
       const publishCmd = `cd ${filename} && npm publish`
+
       exec(publishCmd, (err, stdout, stderr) => {
         if (err) {
-          console.log(chalk.red('Publish Fail ... ' + err))
+          logger.fail('Publish Fail ... ' + err)
           process.exit()
         } else {
-          console.log(chalk.green('Publish Success ...'))
+          logger.success('Publish Success ...')
           process.exit()
         }
       });
-
-      // process.exit()
-
 
       // todo
       // 增加一个参数 --git, 创建一个github仓库
